Updated folder structure (#42)
* chore: updated folder structure

* chore: add auth file
