--- src/domain/daemon/mod.rs
@@ -1,2 +1,0 @@
-pub mod namespace;
-pub mod port;
