--- src/domain/management/mod.rs
@@ -1,3 +1,0 @@
-pub mod port;
-pub mod project;
-pub mod user;
