+++ src/domain/mod.rs
@@ -1,3 +1,4 @@
-pub mod daemon;
-pub mod management;
 pub mod events;
+pub mod ports;
+pub mod projects;
+pub mod users;
