--- src/driven/legacy/account.rs
@@ -1,1 +1,0 @@
-
